fix sdk import issues
