+++ packages/libs/src/sdk/api/comments/CommentsAPI.ts
@@ -12,14 +12,8 @@ import {
   CommentsApi as GeneratedCommentsApi
 } from '../generated/default'
 
-export type CommentMetadata = {
-  body?: string
-  userId: number
-  entityId: number
-  entityType?: EntityType // For now just tracks are supported, but we left the door open for more
-  parentCommentId?: number
-  timestamp_s?: number
-}
+import { CommentMetadata } from './types'
+
 export class CommentsApi extends GeneratedCommentsApi {
   constructor(
     configuration: Configuration,
